+++ projects/ontimize-web-ngx/src/lib/components/o-service-component.class.ts
@@ -1,7 +1,7 @@
 import { SelectionModel } from '@angular/cdk/collections';
 import { ElementRef, EventEmitter, forwardRef, Injector, NgZone, ViewChild } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { BehaviorSubject, Observable } from 'rxjs';
+import { BehaviorSubject, Observable, Subscription } from 'rxjs';
 
 import { OFilterBuilderComponent } from '../components/filter-builder/o-filter-builder.component';
 import { OSearchInputComponent } from '../components/input/search-input/o-search-input.component';
@@ -89,6 +89,14 @@ export const DEFAULT_INPUTS_O_SERVICE_COMPONENT = [
   'quickFilterPlaceholder: quick-filter-placeholder',
 ];
 
+export const DEFAULT_OUTPUTS_O_SERVICE_COMPONENT = [
+  'onClick',
+  'onDoubleClick',
+  'onDataLoaded',
+  'onPaginatedDataLoaded',
+  'onSearch'
+]
+
 export abstract class AbstractOServiceComponent<T extends AbstractComponentStateService<AbstractComponentStateClass>>
   extends AbstractOServiceBaseComponent<T>
   implements IServiceDataComponent {
@@ -174,8 +182,11 @@ export abstract class AbstractOServiceComponent<T extends AbstractComponentState
   /* end of inputs variables */
 
   /* outputs variables */
+  public onClick: EventEmitter<any> = new EventEmitter();
+  public onDoubleClick: EventEmitter<any> = new EventEmitter();
   public onDataLoaded: EventEmitter<any> = new EventEmitter();
   public onPaginatedDataLoaded: EventEmitter<any> = new EventEmitter();
+  public onSearch: EventEmitter<any> = new EventEmitter();
   /* end of outputs variables */
 
   public filterBuilder: OFilterBuilderComponent;
@@ -194,6 +205,7 @@ export abstract class AbstractOServiceComponent<T extends AbstractComponentState
   protected quickFilterColArray: string[];
 
   protected dataResponseArray: any[] = [];
+  protected quickFilterSubscription: Subscription;
 
   constructor(
     injector: Injector,
@@ -245,6 +257,9 @@ export abstract class AbstractOServiceComponent<T extends AbstractComponentState
     if (this.tabsSubscriptions) {
       this.tabsSubscriptions.unsubscribe();
     }
+    if (this.quickFilterSubscription) {
+      this.quickFilterSubscription.unsubscribe();
+    }
   }
 
   public isVisible(): boolean {
@@ -268,7 +283,9 @@ export abstract class AbstractOServiceComponent<T extends AbstractComponentState
   }
 
   public setSelected(item: any): void {
-    this.selection.toggle(item);
+    if (Util.isDefined(item)) {
+      this.selection.toggle(item);
+    }
   }
 
   protected navigateToDetail(route: any[], qParams: any, relativeTo: ActivatedRoute): void {
@@ -573,12 +590,17 @@ export abstract class AbstractOServiceComponent<T extends AbstractComponentState
       return;
     }
     this.quickFilterComponent = quickFilter;
-    if (Util.isDefined(this.quickFilterComponent) && Util.isDefined(this.state)) {
-      this.quickFilterComponent.onSearch.subscribe(val => this.filterData(val));
-      if ((this.state.quickFilterValue || '').length > 0) {
-        this.quickFilterComponent.setValue(this.state.quickFilterValue, {
-          emitEvent: true
-        });
+    if (Util.isDefined(this.quickFilterComponent)) {
+      this.quickFilterSubscription = this.quickFilterComponent.onSearch.subscribe(val => {
+        this.onSearch.emit(val);
+        this.filterData(val);
+      });
+      if (Util.isDefined(this.state)) {
+        if ((this.state.quickFilterValue || '').length > 0) {
+          this.quickFilterComponent.setValue(this.state.quickFilterValue, {
+            emitEvent: true
+          });
+        }
       }
     }
   }
@@ -681,9 +703,6 @@ export abstract class AbstractOServiceComponent<T extends AbstractComponentState
         this.filterData();
       }
     }
-    if (this.loaderSubscription) {
-      this.loaderSubscription.unsubscribe();
-    }
     if (this.pageable) {
       ObservableWrapper.callEmit(this.onPaginatedDataLoaded, data);
     }
@@ -710,7 +729,7 @@ export abstract class AbstractOServiceComponent<T extends AbstractComponentState
             updateComponentStateSubject.next(arg);
           }
         }
-        this.checkViewPortSubject.next(true)
+        this.checkViewPortSubject.next(true);
       });
 
       this.tabsSubscriptions.add(updateComponentStateSubject.subscribe((arg) => {